refactor: dto 패키지 이동
- presentation 계층으로 이동
