+++ src/main/java/roomescape/theme/presentation/ThemeController.java
@@ -10,9 +10,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.theme.dto.PopularThemeResponse;
-import roomescape.theme.dto.ThemeRequest;
-import roomescape.theme.dto.ThemeResponse;
+import roomescape.theme.presentation.dto.PopularThemeResponse;
+import roomescape.theme.presentation.dto.ThemeRequest;
+import roomescape.theme.presentation.dto.ThemeResponse;
 import roomescape.theme.service.ThemeService;
 
 @RestController
