+++ src/main/java/roomescape/theme/presentation/dto/ThemeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.theme.dto;
+package roomescape.theme.presentation.dto;
 
 import roomescape.theme.domain.Theme;
