+++ src/main/java/roomescape/theme/service/ThemeService.java
@@ -7,9 +7,9 @@ import roomescape.common.util.DateTime;
 import roomescape.reservation.domain.ReservationRepository;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.domain.ThemeRepository;
-import roomescape.theme.dto.PopularThemeResponse;
-import roomescape.theme.dto.ThemeRequest;
-import roomescape.theme.dto.ThemeResponse;
+import roomescape.theme.presentation.dto.PopularThemeResponse;
+import roomescape.theme.presentation.dto.ThemeRequest;
+import roomescape.theme.presentation.dto.ThemeResponse;
 
 @Service
 public class ThemeService {
