+++ src/test/java/roomescape/theme/ThemeIntegrationTest.java
@@ -15,7 +15,7 @@ import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
 import roomescape.common.exception.handler.dto.ExceptionResponse;
-import roomescape.theme.dto.ThemeResponse;
+import roomescape.theme.presentation.dto.ThemeResponse;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
